BLAST 2 Folders

BLAST 2 Folders with FASTA files containing nucleotide or protein sequences with a QuickBLAST instance. The files from query and subject folders are selected with the extension parameter

BLAST2Folders(
  ptr,
  query,
  subject,
  extension,
  out_folder,
  out_format = NULL,
  num_threads = 0L,
  reciprocal_hits = FALSE,
  min_batch_size = 0L,
  verbose = TRUE
)

Arguments

ptr

(Rcpp::XPtr<QuickBLAST>) or (unsigned int) Pointer/ID of QuickBLAST instance

query

(string) Query folder

subject

(string) Subject folder.

extension

(string) Extension of files in folder.

out_folder

(string) Ouput Folder (Required).

out_format

(string) Ouput Format. 'ipc'/'csv'/'parquet' (Optional) (Default: 'parquet').

num_threads

(unsigned int) Number of threads. (Optional)

reciprocal_hits

(bool) Perform Bi-directional (Reciprocal => query <-> subject) BLAST? (Default: FALSE) (Optional)

min_batch_size

(unsigned int) Minimum batch size - Size of file write buffer (Optional).

verbose

(bool) Verbosity (Defaut: TRUE).

Value

(bool) TRUE - on success, FALSE - Otherwise. (Results are not returned as R Lists to reduce overhead)

Note

Only FASTA files are supported by this function, use BLAST2DBs() if inputs are BLAST DBs.

See also